Use the default ffmpeg Nix package
The ffmpeg 5 package has been removed in newer versions of NixOS.
This results in an error when attempting to override nixpkgs to a newer version: `error: ffmpeg_5-headless has been removed, please use another version`
Using the default ffmpeg-headless package makes it easier to use a different version of ffmpeg / nixpkgs when incorporating this flake in another.
This PR just changes the flake to use the default ffmpeg package.
The flake isn't updated, so it is still using ffmpeg version 5.